Notes for developes and contributors

Mental classification of functions per functionality

(must clarify these notes) To expand upon: allocate - deallocate reserve - unreserve function getters get... -> get field new... -> construct something new prioritize -> return a vector of indices choose -> return a specific (scalar) index comp -> compilation

How to read variables

transmission -> trans
module -> mdl
mode -> mode
index -> idx
uuid -> id
ibn framework -> ibnf

Testing

Usefull testing functions are in the TestModule weak dependency. To access them you first need to load Test and JET and then use Base.get_extention.

# get the test module from MINDFul
import Test, JET
TestModule = Base.get_extension(MINDFul, :TestModule)
@test !isnothing(TM)

Now, with dot notation (TestModule.) you can access all the following functions.

Testing API

TestModule.JETfilteroutfunctions — Method

JETfilteroutfunctions(f) -> Bool

TestModule.islowlevelintentdagnodeinstalled — Method

islowlevelintentdagnodeinstalled(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    lli::MINDFul.LowLevelIntent
) -> Union{Nothing, Test.Error, Test.Fail, Test.Pass}

TestModule.localnodesaregraphnodeidx — Method

localnodesaregraphnodeidx(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView
)

TestModule.nothingisallocated — Method

nothingisallocated(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView
)

TestModule.testcompilation — Method

testcompilation(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    withremote
) -> Union{Nothing, Test.Error, Test.Fail, Test.Pass}

TestModule.testinstallation — Method

testinstallation(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    withremote
) -> Union{Nothing, Test.Error, Test.Fail, Test.Pass}

TestModule.testlocalnodeisindex — Method

testlocalnodeisindex(
    ibnf
) -> Union{Test.Error, Test.Fail, Test.Pass}

TestModule.testoxcfiberallocationconsistency — Method

testoxcfiberallocationconsistency(ibnf)

Check if the IBNFramework fiber allocations are done from both endpoint oxcviews

TestModule.testuncompilation — Method

testuncompilation(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID
) -> Union{Test.Error, Test.Fail, Test.Pass}

TestModule.testuninstallation — Method

testuninstallation(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    withremote,
    shouldempty
)