MINDFul.KSPFFalg — Constant

The keyword for KShorestPathFirstFitCompilation

MINDFul.AbstractIBNFHandler — Type

A handler or API for IBNFrameworks to talk to each other

MINDFul.AbstractIntent — Type

abstract type AbstractIntent

A concrete subtype of AbstractIntent must implement the following methods:

is_low_level_intent

MINDFul.AbstractIntentConstraint — Type

Special requirements for an intent (such as QoS)

MINDFul.AbstractOperationMode — Type

How the IBN frameworks operates generally. It has effect of the available compilation algorithms In the future it could also have effect on different intent state machines

MINDFul.AbstractSDNController — Type

The abstract type of the SDN controllers

MINDFul.ConnectivityIntent — Type

struct ConnectivityIntent{T<:MINDFul.AbstractIntentConstraint} <: MINDFul.AbstractIntent

sourcenode::MINDFul.GlobalNode: Source node

destinationnode::MINDFul.GlobalNode: Destination node

rate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}: Bandwidth request value (Gbps)

constraints::Vector{T} where T<:MINDFul.AbstractIntentConstraint: Constraints for the Connecivity intent

MINDFul.DefaultOperationMode — Type

Default operation mode defined in MINDFul.jl

MINDFul.EdgeProperties — Type

struct EdgeProperties

distance::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}: The distance of the edge (assumed km)

An immutable description of the edge properties

MINDFul.EdgeView — Type

struct EdgeView

edgeproperties::MINDFul.EdgeProperties: The EdgeProperties

The view of the current edge settings

MINDFul.GlobalNode — Type

struct GlobalNode

ibnfid::Base.UUID: IBN Framework id

localnode::Int64: Node number

MINDFul.IBNAttributeGraph — Type

The graph of the IBN Framework is expressed with this `AttributeGraph`.
Border nodes are assumed to be visible from both sides.
However only the official owner can issue an intent.

MINDFul.IBNFramework — Type

struct IBNFramework{O<:MINDFul.AbstractOperationMode, S<:MINDFul.AbstractSDNController, T<:(AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView), H<:MINDFul.AbstractIBNFHandler} <: MINDFul.AbstractIBNFHandler

operationmode::MINDFul.AbstractOperationMode: The operation mode of the IBN framework

ibnfid::Base.UUID: The id of this IBN Framework instance

intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}: The intent dag tree that contains all intents (can be disconnected graph)

ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView: Single-domain internal graph with border nodes included

ibnfhandlers::Vector{H} where H<:MINDFul.AbstractIBNFHandler: Other IBN Frameworks handles

sdncontroller::MINDFul.AbstractSDNController: SDN controller handle

MINDFul.IBNFramework — Method

IBNFramework(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView
) -> MINDFul.IBNFramework{MINDFul.DefaultOperationMode, MINDFul.SDNdummy, T, H} where {T<:(AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView), H<:(MINDFul.IBNFramework{MINDFul.DefaultOperationMode, MINDFul.SDNdummy})}

The most default construct with abstract type of IBN handlers

MINDFul.IBNFramework — Method

IBNFramework(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    ibnfhandlers::Array{H<:MINDFul.AbstractIBNFHandler, 1}
) -> MINDFul.IBNFramework{MINDFul.DefaultOperationMode, MINDFul.SDNdummy}

Constructor that specify IBNFHandlers to make it potentially type stable

MINDFul.IBNIssuer — Type

struct IBNIssuer <: MINDFul.IntentIssuer

Intent is issued by an IBN Framework domain

ibnfid::Base.UUID: the id of the IBNF issued the intent

idagnodeid::Base.UUID: The id of the intent node in the DAG. The issuer of this intent node points back in this IBNIssuer instance.

MINDFul.IntentCompilationAlgorithm — Type

Instances of this specify how to compile the intent

MINDFul.IntentDAGInfo — Method

IntentDAGInfo() -> MINDFul.IntentDAGInfo

Empty constructor

MINDFul.IntentDAGNode — Type

struct IntentDAGNode{I<:MINDFul.AbstractIntent, II<:MINDFul.IntentIssuer}

intent::MINDFul.AbstractIntent: The intent itself

idagnodeid::Base.UUID: The id of the intent w.r.t. the intent DAG it belongs

intentissuer::MINDFul.IntentIssuer: The intent issuer

logstate::Vector{Tuple{Dates.DateTime, MINDFul.IntentState.T}}: The history of states of the intent with the last being the current state

MINDFul.IntentIssuer — Type

Characterizes the entity issuing an intent

MINDFul.IntentLogState — Type

IntentLogState(

) -> Vector{Tuple{Dates.DateTime, MINDFul.IntentState.T}}
IntentLogState(
    intentstate::MINDFul.IntentState.T
) -> Vector{Tuple{Dates.DateTime, MINDFul.IntentState.T}}

MINDFul.IntentLogState — Type

Stores a vector of the history of the intent states and their timings

MINDFul.KShorestPathFirstFitCompilation — Type

struct KShorestPathFirstFitCompilation <: MINDFul.IntentCompilationAlgorithm

k::Int64: How many k paths to check

MINDFul.LightpathIntent — Type

struct LightpathIntent <: MINDFul.AbstractIntent

path::Vector{Int64}

Expresses an intent for a lightpath. Compilation should yield:

source and destination port indices

transmissionmodule selection

MINDFul.LocalNode — Type

Signifies a local node notation as a single integer

MINDFul.LowLevelIntent — Type

abstract type LowLevelIntent <: MINDFul.AbstractIntent

An abstract subtype of AbstractIntent notating device-level intents and should return is_low_level_intent to be true

MINDFul.MachineGenerated — Type

Intent is generated automatically by the IBN Framework

MINDFul.NetworkOperator — Type

Intent issued directly by the network operator, i.e., a user intent

MINDFul.NodeProperties — Type

struct NodeProperties

localnode::Int64

globalnode::MINDFul.GlobalNode

latitude::Float64

longitude::Float64

inneighbors::Vector{Int64}: The list of neighbohrs coming in

outneighbors::Vector{Int64}: The list of neighbohrs going out

An immutable description of the node properties

MINDFul.NodeView — Type

struct NodeView{R<:MINDFul.RouterView, O<:MINDFul.OXCView, T<:MINDFul.TransmissionModuleView} <: MINDFul.ReservableResourceView

nodeproperties::MINDFul.NodeProperties: The NodeProperties

routerview::Union{Nothing, R} where R<:MINDFul.RouterView: The router in use

oxcview::Union{Nothing, O} where O<:MINDFul.OXCView: The OXC in use

transmissionmoduleviewpool::Union{Nothing, Vector{T}} where T<:MINDFul.TransmissionModuleView: The transmission modules contained

transmissionmodulereservations::Union{Nothing, Dict{Base.UUID, MINDFul.TransmissionModuleLLI}}: intent reservation of the transmission modules

The view of the current node settings Reservables resources reservations are nothing if it represents a border node.

MINDFul.OXCAddDropBypassSpectrumLLI — Type

struct OXCAddDropBypassSpectrumLLI <: MINDFul.LowLevelIntent

localnode::Int64: Target node

localnode_input::Int64: The node in graph entering the OXC (or 0 if invalid)

adddropport::Int64: The port index adding or dropping an optical signal (or 0 if invalid)

localnode_output::Int64: The node in graph exiting the OXC (or 0 if invalid)

spectrumslotsrange::UnitRange{Int64}: The spectrum range allocated 1-based indexed

A value-based interpretation of (input, adddrop, output). At least one of the 3 elements must be 0. (x, 0, y) means optical bypass from the localnode x to the localnode y (0, x, y) means adding an optical signal from add port x going to the localnode y (x, y, 0) means droping an optical signal from the localnode x to the drop port y (0, x, 0) mean that an add/drop allocation port is only reserved (is needed on top for an add/drop signal)

MINDFul.OXCView — Type

struct OXCView{O<:MINDFul.AbstractOXC} <: MINDFul.ReservableResourceView

oxc::MINDFul.AbstractOXC: the underlying OXC

adddropportnumber::Int64: The number of add/drop ports in OXC

switchreservations::Dict{Base.UUID, MINDFul.OXCAddDropBypassSpectrumLLI}: The intent reservations together with the configuration

linkspectrumavailabilities::Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, Vector{Bool}}: Link spectrum availability total view in sync with switchreservations. A vector showing the availability of the spectrum slots. true for available and false for reserved. The vector views should be the same with the ones on the opposite OXC view.

A view of a OXC .

MINDFul.OpticalInitiateConstraint — Type

struct OpticalInitiateConstraint <: MINDFul.AbstractIntentConstraint

Constraint that requires the intent to initiate optically. It's combined with an (@ref)[OpticalTerminateConstraint] before. It contains some requirements for the connection to work out.

globalnode_input::MINDFul.GlobalNode: The incoming border node entering the OXC

spectrumslotsrange::UnitRange{Int64}: The spectrum range allocated 1-based indexed

opticalreach::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}: the remaining optical reach to use

transmissionmodulecompat::MINDFul.TransmissionModuleCompatibility: Requirements for termination in the electical layer

MINDFul.OpticalTerminateConstraint — Type

struct OpticalTerminateConstraint <: MINDFul.AbstractIntentConstraint

Constraint that requires the intent to terminate optically one node before the destination. It's combined with an (@ref)[OpticalInitiateConstraint] after.

MINDFul.RemoteIBNFHandler — Type

struct RemoteIBNFHandler <: MINDFul.AbstractIBNFHandler

Fabian Gobantes implementation. Should consist of basic information all handlers should have (e.g. ibnfid). And a parametric type specific to the protocol used.

    struct HandlerProperties
        ibnfid::UUID
        # ...
    end

    struct IBNFHTTP2Comm <: AbstractIBNFComm
        # example
    end

    struct IBNFSameProcess{T<:IBNFramework} <: AbstractIBNFComm
        # this can  be the new dummy and substitute the current dummy implementation
        ibng::T
    end

    struct RemoteIBNFHandler{T<:AbstractIBNFComm} <: AbstractIBNFHandler
        handlerproperties::HandlerProperties
        ibnfcomm::T
    end

MINDFul.RemoteIntent — Type

struct RemoteIntent{I<:MINDFul.AbstractIntent} <: MINDFul.AbstractIntent

ibnfid::Base.UUID: The id of the remote IBN framework

idagnodeid::Base.UUID: The dag node id of the remote IBN framework

intent::MINDFul.AbstractIntent: The intent to be transferred

isinitiator::Bool: true if the intent originates here and false otherwise

MINDFul.ReservableResourceView — Type

The following functions should be implemented for subtypes:

getreservations(subtype::ReservableResourceView)::Dict{UUID, T}

canreserve(subtype::ReservableResourceView, reservation::T)::Bool

The following default functions exist that should already work

reserve!(subtype::ReservableResourceView, dagnodeid::UUID, reservation::T; checkfirst::Bool=true)::Bool

unreserve!(subtype::ReservableResourceView, dagnodeid::UUID)::Bool

insertreservation!(subtype::ReservableResourceView, dagnodeid::UUID, reservation::T)

deletereservation!(subtype::ReservableResourceView, dagnodeid::UUID)

MINDFul.RouterPortLLI — Type

struct RouterPortLLI <: MINDFul.LowLevelIntent

localnode::Int64: Target node

routerportindex::Int64: The router port index to be reserved

MINDFul.RouterView — Type

struct RouterView{R<:MINDFul.AbstractRouter} <: MINDFul.ReservableResourceView

A view of a router with several ports.

router::MINDFul.AbstractRouter: The underlying router

portnumber::Int64: number of ports in router

portreservations::Dict{Base.UUID, MINDFul.RouterPortLLI}: The intent reservations together with the low level intent of reserved port

MINDFul.SDNdummy — Type

struct SDNdummy <: MINDFul.AbstractSDNController

MINDFul.TransmissionMode — Type

struct TransmissionMode

Represents a transmission mode. A transponder, if flexible, might support many of them.

opticalreach::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}: Optical reach in kilometers

rate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}: rate in Gbps

spectrumslotsneeded::Int64: Number of 12.5 GHz frequency slots needed

MINDFul.TransmissionModuleCompatibility — Type

struct TransmissionModuleCompatibility

rate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}: The bandwidth of the connection

spectrumslotsneeded::Int64: Number of 12.5Gz frequency slots needed

name::String: descriptive name of the transmission module

Describe what requirements are needed to establish transmission module end-to-end connection Modulation should be here but it's not. Dissagregetation data should be here but it's not. Both are covered by the name which must be equal in both ends

MINDFul.TransmissionModuleIntent — Type

struct TransmissionModuleIntent <: MINDFul.AbstractIntent

Contains the requirements to compile down to TransmissionModuleLLI

rate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}: The data rate requierement

opticalreach::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}: Optical reach requirements in kilometres

spectrumslotsneeded::Int64: Number of 12.5 GHz frequency slots needed

MINDFul.TransmissionModuleLLI — Type

struct TransmissionModuleLLI <: MINDFul.LowLevelIntent

localnode::Int64: Target node

transmissionmoduleviewpoolindex::Int64: The index of the transmission module pool to be reserved

transmissionmodesindex::Int64: The selected mode of the transmission module. 0 means nothing is still selected. Non elastic modules can have only 1.

MINDFul.TransmissionModuleView — Type

struct TransmissionModuleView{T<:MINDFul.AbstractTransmissionModule}

transmissionmodule::MINDFul.AbstractTransmissionModule: The underlying transmission module

name::String: descriptive name of the transmission module

transmissionmodes::Vector{MINDFul.TransmissionMode}: operating transmission modes

cost::Float64: Cost of the transmission module (in unit costs)

A view of a transmission module.

Base.show — Method

show(io::IO, ibnf::MINDFul.IBNFramework)

MINDFul.addidagnode! — Method

addidagnode!(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    intent::MINDFul.AbstractIntent;
    parentid,
    intentissuer
) -> MINDFul.IntentDAGNode{I, MINDFul.MachineGenerated} where I<:MINDFul.AbstractIntent

Return the IntentDAGNode

MINDFul.addidagnode! — Method

addidagnode!(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnode::MINDFul.IntentDAGNode;
    parentid,
    intentissuer
) -> Base.UUID

Return the UUID

MINDFul.addintent! — Method

addintent!(
    ibnf::MINDFul.IBNFramework,
    intent::MINDFul.AbstractIntent,
    intentissuer::MINDFul.IntentIssuer
) -> Base.UUID

Add a new user intent to the IBN framework.

MINDFul.allocate_oxcspectrumslots — Method

allocate_oxcspectrumslots(
    ibnf::MINDFul.IBNFramework,
    dagnodeid::Base.UUID,
    localnode::Int64
) -> Bool

MINDFul.allocate_transmissionmodule — Method

allocate_transmissionmodule(
    ibnf::MINDFul.IBNFramework,
    dagnodeid::Base.UUID,
    localnode::Int64,
    transmissionmodulereservationentry::MINDFul.TransmissionModuleLLI
)

MINDFul.aretransmissionmodulescompatible — Method

aretransmissionmodulescompatible(
    tmv1::MINDFul.TransmissionModuleView,
    tmv2::MINDFul.TransmissionModuleView
) -> Bool

MINDFul.canreserve — Method

canreserve(
    nodeview::MINDFul.NodeView,
    transmissionmodulelli::MINDFul.TransmissionModuleLLI;
    verbose
) -> Bool

Set verbose=true to see where the reservation fails

MINDFul.canreserve — Method

canreserve(
    oxcview::MINDFul.OXCView,
    oxcswitchreservationentry::MINDFul.OXCAddDropBypassSpectrumLLI;
    verbose
) -> Bool

Check whether

add/drop port exists

add/drop port already in use

spectrum in fibers in use

Set verbose=true to see where the reservation fails

MINDFul.canreserve — Method

canreserve(
    routerview::MINDFul.RouterView,
    routerportlli::MINDFul.RouterPortLLI;
    verbose
) -> Bool

Check if router port exists and whether it is already used

Set verbose=true to see where the reservation fails

MINDFul.compileintent! — Method

compileintent!(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    algorithm::MINDFul.IntentCompilationAlgorithm;
    verbose
) -> Bool

MINDFul.compileintent! — Method

compileintent!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.ConnectivityIntent},
    kspffalg::MINDFul.KShorestPathFirstFitCompilation
) -> Bool

MINDFul.compileintent! — Method

compileintent!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.RemoteIntent},
    algorithm::MINDFul.IntentCompilationAlgorithm
) -> Bool

MINDFul.constructfromdict — Method

constructfromdict(
    _::Type{T},
    dict::Dict{Symbol},
    dict2::Dict{Symbol}
) -> MINDFul.NodeProperties

Convenience function to construct any struct of type T by passing a Dict corresponding to the fields needed. A further dict dict2 is searched for data.

MINDFul.constructfromdict — Method

constructfromdict(
    _::Type{T},
    dict::Dict{Symbol}
) -> MINDFul.EdgeProperties

Convenience function to construct any struct of type T by passing a Dict corresponding to the fields needed.

MINDFul.deallocate_transmissionmodule — Method

deallocate_transmissionmodule(
    ibnf::MINDFul.IBNFramework,
    localnode::Int64,
    transmissionmodulereservationentry::MINDFul.TransmissionModuleLLI
) -> Bool

MINDFul.default_OXCview — Method

default_OXCview(
    nodeproperties::MINDFul.NodeProperties,
    spectrumslots::Int64
) -> MINDFul.OXCView{MINDFul.OXCDummy}

MINDFul.deletereservation! — Method

deletereservation!(
    resourceview::MINDFul.ReservableResourceView,
    dagnodeid::Base.UUID;
    verbose
)

MINDFul.deletereservationhook! — Method

deletereservationhook!(
    oxcview::MINDFul.OXCView,
    dagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.deletereservationhook! — Method

deletereservationhook!(
    resourceview::MINDFul.ReservableResourceView,
    dagnodeid::Base.UUID;
    verbose
) -> Bool

Implement this function to do custom actions per specific ReservableResourceView

MINDFul.displayavailablecompilationalgorithmsinfo — Method

displayavailablecompilationalgorithmsinfo(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler
)

MINDFul.edgeify — Method

edgeify(
    path::Vector{Int64}
) -> Vector{Graphs.SimpleGraphs.SimpleEdge{Int64}}

MINDFul.findindexglobalnode — Method

findindexglobalnode(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    globalnode::MINDFul.GlobalNode
) -> Union{Nothing, Int64}

Convenience function that returns the findfirst for the global node

MINDFul.firstfit — Method

firstfit(
    boolvec::AbstractVector{Bool},
    lenghrequire::Int64
) -> Union{Nothing, Int64}

Finds first contiguous slot range of length lengthrequire that satisfies the boolvec. Return the starting index of the range or nothing if none available

MINDFul.generatelightpathoxcadddropbypassspectrumlli — Method

generatelightpathoxcadddropbypassspectrumlli(
    path::Vector{Int64},
    spectrumslotsrange::UnitRange{Int64};
    sourceadddropport,
    opticalinitincomingnode,
    destadddropport
) -> Vector{MINDFul.OXCAddDropBypassSpectrumLLI}

Return a list of (@ref)[OXCAddDropBypassSpectrumLLI] that constitute a lightpath. According to the arguments a "starting lightpath", an "ending lightpath", or a "lightpath segment" can be created. The arguments are the following:

sourceadddropport: nothing for a starting lightpath and an Integer denating the add/drop port otherwise

opticalinitincomingnode : the (@ref)[LocalNode] denoting the incoming optical connection from the specified node for a starting lightpath. Set to nothing if not a starting lightpath

destadddropport: nothing for an ending lightpath and an Integer denating the add/drop port otherwise

Note: not both sourceadddropport and opticalinitincomingnode can be nothing or have a values at the same time.

MINDFul.getadddropportnumber — Method

getadddropportnumber(oxcview::MINDFul.OXCView) -> Int64

MINDFul.getafterlliidx — Method

getafterlliidx(
    ibnf::MINDFul.IBNFramework,
    conintent::MINDFul.ConnectivityIntent,
    llis,
    oxclli::MINDFul.OXCAddDropBypassSpectrumLLI;
    verbose
) -> Any

Return all next logical low level intent indices from llis given that now signal is positioned in OXCAddDropBypassSpectrumLLI such that the conintent is satisfied. Return an empty vector if no logical next is found.

MINDFul.getafterlliidx — Method

getafterlliidx(
    ibnf::MINDFul.IBNFramework,
    conintent::MINDFul.ConnectivityIntent,
    llis,
    rplli::MINDFul.RouterPortLLI;
    verbose
) -> Any

Return all next logical low level intent indices from llis given that now signal is positioned in RouterPortLLI such that the conintent is satisfied. Return an empty vector if no logical next is found.

MINDFul.getafterlliidx — Method

getafterlliidx(
    ibnf::MINDFul.IBNFramework,
    conintent::MINDFul.ConnectivityIntent,
    llis,
    tmlli::MINDFul.TransmissionModuleLLI;
    verbose
) -> Any

Return all next logical low level intent indices from llis given that now signal is positioned in TransmissionModuleLLI such that the conintent is satisfied. Return an empty vector if no logical next is found.

MINDFul.getavailabletransmissionmoduleviewindex — Method

getavailabletransmissionmoduleviewindex(
    nodeview::MINDFul.NodeView
) -> Vector{Int64}

MINDFul.getborderedges — Method

getborderedges(
    ibnf::MINDFul.IBNFramework
) -> Vector{Graphs.SimpleGraphs.SimpleEdge{Int64}}

Return all border edges that contain at least one border node as endpoints

MINDFul.getborderglobaledges — Method

getborderglobaledges(ibnf::MINDFul.IBNFramework) -> Vector

Return all border edges that contain at least one border node as endpoints as global

MINDFul.getbordernodesasglobal — Method

getbordernodesasglobal(ibnf::MINDFul.IBNFramework) -> Vector

Return all border nodes of ibnf with globalnode representation

MINDFul.getbordernodesaslocal — Method

getbordernodesaslocal(ibnf::MINDFul.IBNFramework) -> Vector

Return all border nodes of ibnf with localnode representation

MINDFul.getcompilationalgorithmtype — Method

getcompilationalgorithmtype(
    s::Val{:kspff}
) -> Type{MINDFul.KShorestPathFirstFitCompilation}

Give back the algorithm mapped to the symbol

MINDFul.getcost — Method

getcost(s) -> Any

MINDFul.getdefaultcompilationalgorithm — Method

getdefaultcompilationalgorithm(
    ibnff::MINDFul.IBNFramework
) -> Symbol

Can overload for different Operation Modes.

MINDFul.getdefaultcompilationalgorithmargs — Method

getdefaultcompilationalgorithmargs(
    s::Val{:kspff}
) -> Tuple{Int64}

MINDFul.getfiberspectrumavailabilities — Method

getfiberspectrumavailabilities(
    ibnf,
    edge::Graphs.SimpleGraphs.SimpleEdge{Int64};
    checkfirst
) -> Vector{Bool}

Get the spectrum availability slots vector for edge

MINDFul.getfirst — Method

getfirst(predicate::Function, ar::AbstractArray) -> Any

Returns the element is predicate is satisfied or nothing otherwise.

MINDFul.getfirstavailableoxcadddropport — Method

getfirstavailableoxcadddropport(
    nodeview::MINDFul.NodeView
) -> Union{Nothing, Int64}

MINDFul.getfirstavailableoxcadddropport — Method

getfirstavailableoxcadddropport(
    oxcview::MINDFul.OXCView
) -> Union{Nothing, Int64}

Return the first available oxc add/drop port and nothing if none found

MINDFul.getfirstavailablerouterportindex — Method

getfirstavailablerouterportindex(
    nodeview::MINDFul.NodeView
) -> Union{Nothing, Int64}

MINDFul.getfirstavailablerouterportindex — Method

getfirstavailablerouterportindex(
    routerview::MINDFul.RouterView
) -> Union{Nothing, Int64}

Return the first available router port index and nothing if non available.

MINDFul.getfirstcompatibletransmoduleidxandmodeidx — Method

getfirstcompatibletransmoduleidxandmodeidx(
    transmissionmoduleviewpool::Vector{<:MINDFul.TransmissionModuleView},
    availtransmdlidxs::Vector{Int64},
    transmissionmodulecompat::MINDFul.TransmissionModuleCompatibility
) -> Union{Nothing, Tuple{Int64, Int64}}

Return a integer tuple (Int, Int) denoting the first available and compatible transmission module index and its transmission mode index. If non found return nothing.

MINDFul.getglobalnode — Method

getglobalnode(s) -> Any

MINDFul.getglobalnode — Method

getglobalnode(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    localnode::Int64
) -> MINDFul.GlobalNode

Return the global representation given the local representation. Return nothing if not found

MINDFul.getibnfid — Method

getibnfid(ri::MINDFul.RemoteIntent) -> Base.UUID

MINDFul.getidagcounter — Method

getidagcounter(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> Int64

MINDFul.getidaginfo — Method

getidaginfo(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> MINDFul.IntentDAGInfo

MINDFul.getidagnextuuidcounter — Method

getidagnextuuidcounter(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> Base.UUID

MINDFul.getidagnode — Method

getidagnode(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    dagnodeid::Base.UUID
) -> MINDFul.IntentDAGNode

MINDFul.getidagnodechildren — Method

getidagnodechildren(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID
) -> Vector{MINDFul.IntentDAGNode}

MINDFul.getidagnodechildren — Method

getidagnodechildren(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnode::MINDFul.IntentDAGNode
) -> Vector{MINDFul.IntentDAGNode}

MINDFul.getidagnodedescendants — Method

getidagnodedescendants(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID;
    exclusive
) -> Vector{MINDFul.IntentDAGNode}

Get all descendants of DAG dag starting from node idagnodeid Set exclusive=true to get nodes that have idagnodeid as the only ancestor

MINDFul.getidagnodeid — Method

getidagnodeid(ri::MINDFul.RemoteIntent) -> Base.UUID

MINDFul.getidagnodeidx — Method

getidagnodeidx(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    dagnodeid::Base.UUID
) -> Int64

Get the vertex index of the intent DAG node with id dagnodeid. Errors if UUID doesn't exist.

MINDFul.getidagnodeleafs — Method

getidagnodeleafs(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID;
    exclusive
) -> Vector{MINDFul.IntentDAGNode}

Get the leafs of DAG dag starting from node idn. Set exclusive=true to get nodes that have idn as the only ancestor

MINDFul.getidagnodellis — Method

getidagnodellis(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID;
    exclusive
) -> Vector{MINDFul.IntentDAGNode}

Get all the Low Level Intents that are leafs of idagnodeid Set exclusive=true to get nodes that have idn as the only ancestor

MINDFul.getidagnodellis — Method

getidagnodellis(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> Vector{MINDFul.IntentDAGNode}

Get all the Low Level Intents that exist on the DAG

MINDFul.getidagnodeparents — Method

getidagnodeparents(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID
) -> Vector{MINDFul.IntentDAGNode}

MINDFul.getidagnodeparents — Method

getidagnodeparents(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnode::MINDFul.IntentDAGNode
) -> Vector{MINDFul.IntentDAGNode}

MINDFul.getidagnodes — Method

getidagnodes(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> Vector{MINDFul.IntentDAGNode}

MINDFul.getidagnodestate — Method

getidagnodestate(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    dagnodeid::Base.UUID
) -> MINDFul.IntentState.T

MINDFul.getinneighbors — Method

getinneighbors(np::MINDFul.NodeProperties) -> Vector{Int64}

MINDFul.getintent — Method

getintent(
    ri::MINDFul.RemoteIntent
) -> MINDFul.AbstractIntent

MINDFul.getisinitiator — Method

getisinitiator(ri::MINDFul.RemoteIntent) -> Bool

MINDFul.getlatitude — Method

getlatitude(np::MINDFul.NodeProperties) -> Float64

MINDFul.getlinkspectrumavailabilities — Method

getlinkspectrumavailabilities(
    oxcview::MINDFul.OXCView
) -> Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, Vector{Bool}}

MINDFul.getlocalnode — Method

getlocalnode(s) -> Any

MINDFul.getlocalnode — Method

getlocalnode(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    globalnode::MINDFul.GlobalNode
) -> Any

Return the localnode representation given the global representation. Return nothing if not found

MINDFul.getlocalnode_input — Method

getlocalnode_input(
    oxcswitchreservationentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Int64

MINDFul.getlocalnode_output — Method

getlocalnode_output(
    oxcswitchentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Int64

MINDFul.getlongitude — Method

getlongitude(np::MINDFul.NodeProperties) -> Float64

MINDFul.getlowestratetransmissionmode — Method

getlowestratetransmissionmode(
    transmissionmoduleview::MINDFul.TransmissionModuleView,
    demandrate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}},
    demanddistance::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}
) -> Union{Nothing, Int64}

Return the index with the lowest GBPS rate that can get deployed for the given demand rate and distance. If non is find return nothing.

MINDFul.getnodeproperties — Method

getnodeproperties(
    s::MINDFul.NodeView
) -> MINDFul.NodeProperties

MINDFul.getnumberofparameters — Method

getnumberofparameters(
    _::MINDFul.IntentCompilationAlgorithm
) -> Int64

MINDFul.getnumberofparameters — Method

getnumberofparameters(
    _::Type{T<:MINDFul.IntentCompilationAlgorithm}
) -> Int64

MINDFul.getopticalinitiateconstraint — Method

getopticalinitiateconstraint(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID
) -> Union{Nothing, MINDFul.OpticalInitiateConstraint}

Get the OpticalInitiateConstraint for the current intent DAG. If the compilation is not optically terminated return nothing.

To me this has all the logic needed to be type stable but the compiler fails.

MINDFul.getopticalreach — Method

getopticalreach(
    tm::MINDFul.TransmissionMode
) -> Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}

MINDFul.getoutneighbors — Method

getoutneighbors(np::MINDFul.NodeProperties) -> Vector{Int64}

MINDFul.getoxcadddropportindex — Method

getoxcadddropportindex(
    s::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Int64

MINDFul.getoxcview — Method

getoxcview(
    s::MINDFul.NodeView
) -> Union{Nothing, MINDFul.OXCView}

MINDFul.getpathspectrumavailabilities — Method

getpathspectrumavailabilities(
    ibnf::MINDFul.IBNFramework,
    localnodespath::Vector{Int64};
    checkfirst
) -> Any

Get spectrum availabilities along a path of nodes as a BitVector

MINDFul.getport_adddrop — Method

getport_adddrop(
    oxcswitchentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Int64

MINDFul.getproperties — Method

getproperties(s::MINDFul.EdgeView) -> MINDFul.EdgeProperties

MINDFul.getproperties — Method

getproperties(s::MINDFul.NodeView) -> MINDFul.NodeProperties

MINDFul.getrate — Method

getrate(
    tm::MINDFul.TransmissionMode
) -> Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}

MINDFul.getrate — Method

getrate(
    tmc::MINDFul.TransmissionModuleCompatibility
) -> Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}}

MINDFul.getreservations — Method

getreservations(
    nodeview::MINDFul.NodeView
) -> Union{Nothing, Dict{Base.UUID, MINDFul.TransmissionModuleLLI}}

MINDFul.getreservations — Method

getreservations(
    oxcview::MINDFul.OXCView
) -> Dict{Base.UUID, MINDFul.OXCAddDropBypassSpectrumLLI}

MINDFul.getreservations — Method

getreservations(
    routerview::MINDFul.RouterView
) -> Dict{Base.UUID, MINDFul.RouterPortLLI}

MINDFul.getreservedtransmissionmode — Method

getreservedtransmissionmode(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{MINDFul.TransmissionModuleLLI};
    verbose
) -> Union{Bool, MINDFul.TransmissionMode}

Get the reserved transmission mode

MINDFul.getrouterportindex — Method

getrouterportindex(s::MINDFul.RouterPortLLI) -> Int64

MINDFul.getrouterview — Method

getrouterview(
    s::MINDFul.NodeView
) -> Union{Nothing, MINDFul.RouterView}

MINDFul.getspectrumslotsneeded — Method

getspectrumslotsneeded(
    tm::MINDFul.TransmissionMode
) -> Int64

MINDFul.getspectrumslotsneeded — Method

getspectrumslotsneeded(
    tmc::MINDFul.TransmissionModuleCompatibility
) -> Int64

MINDFul.getspectrumslotsrange — Method

getspectrumslotsrange(
    oxcswitchentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> UnitRange{Int64}

MINDFul.gettransmissionmode — Method

gettransmissionmode(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    oxclli::MINDFul.TransmissionModuleLLI
) -> MINDFul.TransmissionMode

MINDFul.gettransmissionmode — Method

gettransmissionmode(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{MINDFul.TransmissionModuleLLI}
) -> MINDFul.TransmissionMode

Get the transmission mode

MINDFul.gettransmissionmode — Method

gettransmissionmode(
    ibnf::MINDFul.IBNFramework,
    intent::MINDFul.TransmissionModuleLLI
) -> MINDFul.TransmissionMode

Get the transmission mode

MINDFul.gettransmissionmode — Method

gettransmissionmode(
    s::MINDFul.TransmissionModuleView,
    transmissionmodeidx::Int64
) -> MINDFul.TransmissionMode

MINDFul.gettransmissionmodes — Method

gettransmissionmodes(
    s::MINDFul.TransmissionModuleView
) -> Vector{MINDFul.TransmissionMode}

MINDFul.gettransmissionmodesindex — Method

gettransmissionmodesindex(
    s::MINDFul.TransmissionModuleLLI
) -> Int64

MINDFul.gettransmissionmodule — Method

gettransmissionmodule(
    ibnag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView,
    oxclli::MINDFul.TransmissionModuleLLI
) -> MINDFul.TransmissionModuleView

MINDFul.gettransmissionmodule — Method

gettransmissionmodule(
    ibnf::MINDFul.IBNFramework,
    intent::MINDFul.TransmissionModuleLLI
) -> MINDFul.TransmissionModuleView

Get the transmission mode

MINDFul.gettransmissionmodulereservations — Method

gettransmissionmodulereservations(
    s::MINDFul.NodeView
) -> Dict{Base.UUID, MINDFul.TransmissionModuleLLI}

MINDFul.gettransmissionmoduleviewpool — Method

gettransmissionmoduleviewpool(
    s::MINDFul.NodeView
) -> Union{Nothing, Vector{T}} where T<:MINDFul.TransmissionModuleView

MINDFul.gettransmissionmoduleviewpoolindex — Method

gettransmissionmoduleviewpoolindex(
    s::MINDFul.TransmissionModuleLLI
) -> Int64

MINDFul.getunderlyingequipment — Method

getunderlyingequipment(
    s::MINDFul.TransmissionModuleView
) -> MINDFul.AbstractTransmissionModule

MINDFul.hasidagnodechildren — Method

hasidagnodechildren(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnodeid::Base.UUID
) -> Bool

MINDFul.hasidagnodechildren — Method

hasidagnodechildren(
    idag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo},
    idagnode::MINDFul.IntentDAGNode
) -> Bool

MINDFul.increaseidagcounter! — Method

increaseidagcounter!(
    intentdag::AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{MINDFul.IntentDAGNode}, Nothing, MINDFul.IntentDAGInfo}
) -> Int64

MINDFul.insertreservation! — Method

insertreservation!(
    resourceview::MINDFul.ReservableResourceView,
    dagnodeid::Base.UUID,
    reservationdescription;
    verbose
) -> Bool

MINDFul.insertreservationhook! — Method

insertreservationhook!(
    oxcview::MINDFul.OXCView,
    dagnodeid::Base.UUID,
    reservationdescription::MINDFul.OXCAddDropBypassSpectrumLLI;
    verbose
) -> Bool

MINDFul.insertreservationhook! — Method

insertreservationhook!(
    resourceview::MINDFul.ReservableResourceView,
    dagnodeid::Base.UUID,
    reservationdescription;
    verbose
) -> Bool

Implement this function to do custom actions per specific ReservableResourceView

MINDFul.installintent! — Method

installintent!(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.is_low_level_intent — Method

is_low_level_intent(ci::MINDFul.ConnectivityIntent) -> Bool

MINDFul.isaddportallocation — Method

isaddportallocation(
    oxcswitchentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Bool

Checks if this reservation reserves the add port, i.e., it's (0, x, y).

MINDFul.isbordernode — Method

isbordernode(
    ibnf::MINDFul.IBNFramework,
    localnode::Int64
) -> Any

Return boolean if localnode is in ibnf as a border node

MINDFul.isbordernode — Method

isbordernode(
    ibnf::MINDFul.IBNFramework,
    globalnode::MINDFul.GlobalNode
) -> Any

Return boolean if globalnode is in ibnf as a border node

MINDFul.isdropportallocation — Method

isdropportallocation(
    oxcswitchentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Bool

Checks if this reservation reserves the drop port, i.e., it's (x, y, 0).

MINDFul.isinternalnode — Method

isinternalnode(
    ibnf::MINDFul.IBNFramework,
    globalnode::MINDFul.GlobalNode
) -> Bool

Return boolean if globalnode belongs to ibnf

MINDFul.isinternalorborderintent — Method

isinternalorborderintent(
    ibnf::MINDFul.IBNFramework,
    connectivityintent::MINDFul.ConnectivityIntent
) -> Bool

Return true if at least source or destination is internal.

MINDFul.isreservationvalid — Function

isreservationvalid(
    oxcswitchreservationentry::MINDFul.OXCAddDropBypassSpectrumLLI
) -> Bool
isreservationvalid(
    oxcswitchreservationentry::MINDFul.OXCAddDropBypassSpectrumLLI,
    verbose::Bool
) -> Bool

MINDFul.issatisfied — Method

issatisfied(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.ConnectivityIntent};
    onlyinstalled,
    noextrallis,
    verbose
) -> Bool

Steps by step check if ibnf satisfies the intent For now works only with local view. The options are:

onlyinstalled: only consideres installed intents

noextrallis: all LLI must be used

orderedllis: pass list to access ordered llis

The function is not a definite assertion. The following cases are not covered:

transmission module compatibility

optical reach

MINDFul.istransmissionmoduleandmodecompatible — Method

istransmissionmoduleandmodecompatible(
    transmissionmoduleview::MINDFul.TransmissionModuleView,
    transmissionmodesindex::Int64,
    transmissionmodulecompat::MINDFul.TransmissionModuleCompatibility
) -> Bool

Return true if the transmissionmoduleview and mode support the transmissionmodulecompat If onlymodecheck = true is passed then the disaggregation/protocol (aka name) will not be checked.

MINDFul.istransmissionmoduleandmodecompatible — Method

istransmissionmoduleandmodecompatible(
    transmissionmoduleview::MINDFul.TransmissionModuleView,
    transmissionmode::MINDFul.TransmissionMode,
    transmissionmodulecompat::MINDFul.TransmissionModuleCompatibility
) -> Bool

Return true if the transmissionmoduleview and mode support the transmissionmodulecompat If onlymodecheck = true is passed then the disaggregation/protocol (aka name) will not be checked.

MINDFul.istransmissionmoduleappropriate — Method

istransmissionmoduleappropriate(
    transmissionmoduleview::MINDFul.TransmissionModuleView,
    demandrate::Unitful.Quantity{Float64, Data 𝐓^-1, Unitful.FreeUnits{(Gbps,), Data 𝐓^-1, nothing}},
    demanddistance::Unitful.Quantity{Float64, 𝐋, Unitful.FreeUnits{(km,), 𝐋, nothing}}
) -> Bool

Checks if the transmission module can get deployed for the given demand rate and distance

MINDFul.istransmissionmodulecompatible — Method

istransmissionmodulecompatible(
    transmissionmoduleview::MINDFul.TransmissionModuleView,
    transmissionmodulecompat::MINDFul.TransmissionModuleCompatibility
) -> Bool

Return true if transmissionmoduleview can support with its modes the transmissionmodulecompat

MINDFul.kspffcrossdomain! — Method

kspffcrossdomain!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.ConnectivityIntent},
    kspffalg::MINDFul.KShorestPathFirstFitCompilation,
    mediatorbordernode::MINDFul.GlobalNode
) -> Bool

MINDFul.kspffintradomain! — Method

kspffintradomain!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.ConnectivityIntent},
    kspffalg::MINDFul.KShorestPathFirstFitCompilation
) -> Bool

MINDFul.kspffintradomain_destination! — Method

kspffintradomain_destination!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode{<:MINDFul.ConnectivityIntent},
    lowlevelintentstoadd,
    transmissionmodulecompat,
    opticalincomingnode::Int64,
    spectrumslotsrange::UnitRange{Int64}
) -> Bool

Takes care of the final node (destination) for the case of no `OpticalTerminateConstraint`

MINDFul.mycopy — Method

mycopy(whatever) -> Any

MINDFul.newoxcentry_adddropallocation — Function

newoxcentry_adddropallocation(port::Int64)
newoxcentry_adddropallocation(
    port::Int64,
    spectrumslotsrange::UnitRange{Int64}
)

MINDFul.pushstatetoidagnode! — Method

pushstatetoidagnode!(
    intentlogstate::Array{Tuple{Dates.DateTime, S}, 1} where S<:Enum{Int32},
    time::Dates.DateTime,
    intentstate::MINDFul.IntentState.T
) -> Array{Tuple{Dates.DateTime, S}, 1} where S<:Enum{Int32}

MINDFul.pushstatetoidagnode! — Method

pushstatetoidagnode!(
    intentlogstate::Array{Tuple{Dates.DateTime, S}, 1} where S<:Enum{Int32},
    intentstate::MINDFul.IntentState.T
) -> Array{Tuple{Dates.DateTime, S}, 1} where S<:Enum{Int32}

Uses now() time as default

MINDFul.pushstatetoidagnode! — Method

pushstatetoidagnode!(
    idagnode::MINDFul.IntentDAGNode,
    time::Dates.DateTime,
    intentstate::MINDFul.IntentState.T
) -> Vector{Tuple{Dates.DateTime, MINDFul.IntentState.T}}

MINDFul.remoteintent! — Method

remoteintent!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode,
    remoteibnfid::Base.UUID
) -> MINDFul.IntentDAGNode{I, MINDFul.MachineGenerated} where I<:MINDFul.RemoteIntent

Add a RemoteIntent as a child intent and delegate it to the ibn with id remoteibndif

MINDFul.removeintent! — Method

removeintent!(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestavailablecompilationalgorithms — Method

requestavailablecompilationalgorithms(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework
) -> Vector{Symbol}

Compilation algorithms are given as symbols because they might not be available programmatically to different IBN frameworks

MINDFul.requestavailablecompilationalgorithms — Method

requestavailablecompilationalgorithms(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler
)

Fabian Gobantes implementation

MINDFul.requestcompileintent_init! — Function

requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID
)
requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol
)
requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol,
    compilationalgorithmargs::Tuple
) -> Bool

The initiator domain myibnf asks remoteibnf to compile the external remote intent idagnodeid with the specified compilation algorithm

MINDFul.requestcompileintent_init! — Function

requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler
)
requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    compilationalgorithm::Symbol
)
requestcompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    compilationalgorithm::Symbol,
    compilationalgorithmkey::Tuple
)

Fabian Gobantes implementation

MINDFul.requestcompileintent_term! — Function

requestcompileintent_term!(
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID
)
requestcompileintent_term!(
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol
)
requestcompileintent_term!(
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol,
    compilationalgorithmargs::Tuple
)

Fabian Gobantes implementation

The initiator domain remoteibnf asks this domain myibnf to compile the internal remote intent idagnodeid with the specified compilation algorithm

MINDFul.requestcompileintent_term! — Function

requestcompileintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID
)
requestcompileintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol
)
requestcompileintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    compilationalgorithmkey::Symbol,
    compilationalgorithmargs::Tuple
)

The initiator domain remoteibnf asks this domain myibnf to compile the internal remote intent idagnodeid with the specified compilation algorithm

MINDFul.requestdelegateintent! — Method

requestdelegateintent!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    intent::MINDFul.AbstractIntent,
    internalidagnodeid::Base.UUID
) -> Base.UUID

Return the id of the new dag node if successful and nothing otherwise

MINDFul.requestdelegateintent! — Method

requestdelegateintent!(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    intent::MINDFul.AbstractIntent,
    internalidagnodeid::Base.UUID
)

Fabian Gobantes implementation

Delegates an intent to another domain

Return the id of the new dag node if successful and nothing otherwise

MINDFul.requestibnattributegraph — Method

requestibnattributegraph(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework
) -> AttributeGraphs.AttributeGraph{Int64, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, Vector{T}, Dict{Graphs.SimpleGraphs.SimpleEdge{Int64}, MINDFul.EdgeView}, Base.UUID} where T<:MINDFul.NodeView

Request topology information

MINDFul.requestibnattributegraph — Method

requestibnattributegraph(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler
)

Fabian Gobantes implementation If far away, think about authorization and permissions. That's the reason why there are 2 arguments: The first argument should have the authorization.

MINDFul.requestinstallintent_init! — Method

requestinstallintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestinstallintent_term! — Method

requestinstallintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestissatisfied — Method

requestissatisfied(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    onlyinstalled,
    noextrallis
)

Request to remoteibnf whether the idagnode is theoretically satisfied

MINDFul.requestissatisfied — Method

requestissatisfied(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.RemoteIBNFHandler,
    idagnode::MINDFul.IntentDAGNode;
    onlyinstalled,
    noextrallis
)

Fabian Gobantes implementation

Request to remoteibnf whether the idagnode is theoretically satisfied

MINDFul.requestremoteintentstateupdate! — Method

requestremoteintentstateupdate!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID,
    newstate::MINDFul.IntentState.T
) -> Bool

Request the initiator remoteibnf to update the state of its mirrored remote intent

MINDFul.requestspectrumavailability — Method

requestspectrumavailability(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    ge::MINDFul.GlobalEdge
) -> Union{Nothing, Vector{Bool}}

Request spectrum slot availabilities of the border edge Need to check whether ge is indeed an edge shared with myibnf

MINDFul.requestspectrumavailability — Method

requestspectrumavailability(
    myibnf::MINDFul.IBNFramework,
    remoteibnfhandler::MINDFul.RemoteIBNFHandler,
    ge::MINDFul.GlobalEdge
)

Fabian Gobantes implementation

Request spectrum slot availabilities of the border edge Need to check whether ge is indeed an edge shared with myibnf

MINDFul.requestuncompileintent_init! — Method

requestuncompileintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestuncompileintent_term! — Method

requestuncompileintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestuninstallintent_init! — Method

requestuninstallintent_init!(
    myibnf::MINDFul.IBNFramework,
    remoteibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.requestuninstallintent_term! — Method

requestuninstallintent_term!(
    remoteibnfhandler::MINDFul.AbstractIBNFHandler,
    myibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.reserve! — Method

reserve!(
    resourceview::MINDFul.ReservableResourceView,
    lowlevelintent::MINDFul.LowLevelIntent,
    dagnodeid::Base.UUID;
    checkfirst,
    verbose
) -> Bool

TODO: put reservations on the OXC edges

MINDFul.reserveunreserveleafintents! — Method

reserveunreserveleafintents!(
    ibnf::MINDFul.IBNFramework,
    idagnodeleaf::MINDFul.IntentDAGNode,
    doinstall::Bool;
    verbose
) -> Bool

to reserve pass doinstall=true, and to unreserve doinstall=false

MINDFul.setoxcviewlinkavailabilities! — Method

setoxcviewlinkavailabilities!(
    oxcview::MINDFul.OXCView,
    oxcadddropbypassspectrumlli::MINDFul.OXCAddDropBypassSpectrumLLI,
    setflag::Bool;
    verbose
) -> Bool

Set the link spectrum availabilities of the oxcview based on the OXC low level intent to setflags

MINDFul.uncompileintent! — Method

uncompileintent!(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.uninstallintent! — Method

uninstallintent!(
    ibnf::MINDFul.IBNFramework,
    idagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.unreserve! — Method

unreserve!(
    resourceview::MINDFul.ReservableResourceView,
    dagnodeid::Base.UUID;
    verbose
) -> Bool

MINDFul.updateidagnodestates! — Method

updateidagnodestates!(
    ibnf::MINDFul.IBNFramework,
    idagnode::MINDFul.IntentDAGNode
) -> Bool

Return value is true if state is changed.

MINDFul.@returniffalse — Macro

Return a return false if the expression ex evaluates to false. If verbose=true print the statement and the location. If the expression passed is true do nothing.

MINDFul.IntentState — Module

All possible default intent states. Another intent state schema could be defined.