API

MINDFul.addintent! — Method

addintent!(ibn::IBN, intent::Intent) -> Int64

Add intent to ibn as Network Operator. Returns the intent id.

MINDFul.addintent! — Method

Add InterIBN-Intent as IBN2IBN, customer2provider

MINDFul.deploy! — Method

ibn-customer accesses the intent state machine of ibn-provider

MINDFul.deploy! — Method

No algmethod provided

MINDFul.deploy! — Method

Assume intent is root of DAG

MINDFul.deploy! — Method

Assume intent comes from Network Operator

MINDFul.edgeify — Method

Converts a node path to a sequence of edges

MINDFul.issatisfied — Method

onlylogic is WIP

MINDFul.issatisfied — Method

check if the path makes sense and the constraints are satisfied

MINDFul.issatisfied — Method

Low Level Intents are assumed to be installed now

MINDFul.issatisfied — Method

Low Level Intents are assumed to be installed now

MINDFul.remintent! — Method

remintent!(ibn::IBN, idi::Int64) -> Bool

Remove intent with id idi from ibn. Returns true if successful.

MINDFul.remintent! — Method

Removes all intents made from this issuer combination

MINDFul.simgraph — Method

Builds a simulated graph Give in a MetaGraph having:

:routerports as integer in every node

:xcoord as integer in every node

:ycoord as integer in every node

:oxc as boolean in every node

:fiberslots as int in every link

Get as an output a MetaGraph having:

:xcoord as integer in every node

:ycoord as integer in every node

:router as RouterView in every node

:link as FiberView in every edge

MINDFul.transnodes — Method

returns trans nodes of the IBN in the format of (IBN id, node id)

MINDFul.ConnectivityIntent — Type

struct ConnectivityIntent{C, R} <: Intent

src::Tuple{Int64, Int64}: Source node as (IBN.id, node-id)

dst::Tuple{Int64, Int64}: Destination node as (IBN.id, node-id)

constraints::Any: Intents constraints

conditions::Any: Intents conditions

Intent for connecting 2 nodes

MINDFul.IBN — Type

struct IBN{T<:SDN}

id::Int64: id of IBN

intents::Vector{MetaGraphsNext.MetaDiGraph{Int64, Base.UUID, Graphs.SimpleGraphs.SimpleDiGraph{Int64}, IntentDAGNode, Nothing, MINDFul.IntentDAGInfo, MetaGraphsNext.var"#3#5", Float64}}: The intent collection of the IBN Framework

intentissuers::Vector{MINDFul.IntentIssuer}

controllers::Array{Union{IBN, T}, 1} where T<:SDN: The collection of SDNs controlled from this IBN Framework and interacting IBNs (future should be IBN-NBIs)

ngr::NestedGraphs.NestedMetaGraph{Int64, MetaGraphs.MetaDiGraph{Int64, Float64}, MetaGraphs.MetaDiGraph{Int64, Float64}}: Nested Graph consisting of the several SDNs ngr is a shallow copy of the sdn graphs, meaning all PHY information is available in the IBN

interprops::Dict{Int64, MINDFul.IBNInterProps}: InterIBN interoperability with key being the IBN id

The Intent Framework The intent id is the vector index controllers must have same length with ngr.grv

MINDFul.IBN — Method

Empty constructor

MINDFul.IntentDAGNode — Type

mutable struct IntentDAGNode{T<:Intent, L<:MINDFul.LogState}

intent::Intent

state::IntentState

id::Base.UUID

logstate::MINDFul.LogState

MINDFul.SDNdummy — Type

A dummy SDN controller to use for research and experiment purposes. This SDN controller is directly connected with the simulated physical layer network resources SimNetResou.jl

gr::MetaGraphs.MetaDiGraph{T} where T: network of SDN

interprops::Dict{NestedGraphs.NestedEdge, Dict{Symbol, Any}}: inter domain equipment(e.g. links)

Index

MINDFul.ConnectivityIntent

MINDFul.IBN

MINDFul.IBN

MINDFul.IntentDAGNode

MINDFul.SDNdummy

MINDFul.addintent!

MINDFul.addintent!

MINDFul.deploy!

MINDFul.deploy!

MINDFul.deploy!

MINDFul.deploy!

MINDFul.edgeify

MINDFul.issatisfied

MINDFul.issatisfied

MINDFul.issatisfied

MINDFul.issatisfied

MINDFul.remintent!

MINDFul.remintent!

MINDFul.simgraph

MINDFul.transnodes